# openapi.yaml
openapi: 3.0.3
info:
  title: HabitatTrack Financial Management API
  version: v1.0.0
  description: API for managing properties, financial transactions, and categories.
  contact:
    name: API Support
    url: http://www.example.com/support
    email: support@example.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.habitattrack.com
    description: Production server
components:
  schemas:
    Property:
      type: object
      required:
        - name
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the property.
          readOnly: true
        name:
          type: string
          description: Name of the property (e.g., "123 Main St Apt 4B").
          example: "Sunrise Apartments Unit 101"
        address:
          type: string
          description: Full address of the property.
          nullable: true
          example: "123 Main St, Anytown, USA 12345"
        # Add other existing property fields if necessary, or keep it minimal as requested
        created_at:
          type: string
          format: date-time
          description: Timestamp of when the property was created.
          readOnly: true
        updated_at:
          type: string
          format: date-time
          description: Timestamp of when the property was last updated.
          readOnly: true
    PropertyCreateRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Name of the property.
        address:
          type: string
          nullable: true
          description: Full address of the property.
    PropertyUpdateRequest:
      type: object
      properties:
        name:
          type: string
          description: Name of the property.
        address:
          type: string
          nullable: true
          description: Full address of the property.
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: An application-specific error code.
          example: "VALIDATION_ERROR"
        message:
          type: string
          description: A human-readable message providing more details about the error.
          example: "The 'name' field is required."
        details:
          type: object
          additionalProperties:
            type: string
          description: Optional field for providing more specific error details, e.g., field-level validation errors.
          example:
            field: "name"
            issue: "cannot be empty"
    PaginationInfo:
      type: object
      properties:
        totalItems:
          type: integer
          format: int64
          description: Total number of items available.
        totalPages:
          type: integer
          format: int32
          description: Total number of pages.
        currentPage:
          type: integer
          format: int32
          description: The current page number.
        pageSize:
          type: integer
          format: int32
          description: Number of items per page.
        nextCursor:
          type: string
          nullable: true
          description: Cursor for fetching the next page (for cursor-based pagination).
        prevCursor:
          type: string
          nullable: true
          description: Cursor for fetching the previous page (for cursor-based pagination).
...
    PaginatedPropertiesResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Property'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
  parameters:
...
    PropertyID:
      name: propertyId
      in: path
      required: true
      description: Identifier of the property.
      schema:
        type: string
        format: uuid
    # Pagination Parameters (Offset/Limit example)
    PageLimit:
      name: limit
      in: query
      description: Maximum number of items to return per page.
      required: false
      schema:
        type: integer
        format: int32
        default: 20
        minimum: 1
        maximum: 100
    PageOffset:
      name: offset
      in: query
      description: Number of items to skip before starting to collect the result set.
      required: false
      schema:
        type: integer
        format: int32
        default: 0
        minimum: 0
    # Pagination Parameters (Cursor-based example)
...
  securitySchemes:
    OAuth2BearerAuth:
      type: oauth2
      description: OAuth 2.0 Bearer Token for API authentication.
      flows:
        implicit: # Or clientCredentials, authorizationCode, password depending on your auth strategy
          authorizationUrl: https://auth.habitattrack.com/oauth2/authorize
          scopes:
            read:transactions: Grants read access to transactions
            write:transactions: Grants write access to transactions
            read:categories: Grants read access to categories
            write:categories: Grants write access to categories
            read:properties: Grants read access to properties
            write:properties: Grants write access to properties
# Global security requirement (can be overridden at operation level)
security:
  - OAuth2BearerAuth: [] # Empty array means no specific scopes required globally, define per operation if needed
paths:
...
  /properties:
    summary: Manage properties.
    get:
      summary: List all properties
      tags: [Properties]
      operationId: listProperties
      parameters:
        - $ref: '#/components/parameters/PageLimit'
        - $ref: '#/components/parameters/PageOffset'
      responses:
        '200':
          description: A paginated list of properties.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedPropertiesResponse'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        default:
          description: Unexpected error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - OAuth2BearerAuth: [read:properties]
    post:
      summary: Create a new property
      tags: [Properties]
      operationId: createProperty
      requestBody:
        description: Property object to be created.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PropertyCreateRequest'
      responses:
        '201':
          description: Property created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
          headers:
            Location:
              description: URL of the newly created property.
              schema:
                type: string
                format: url
        '400':
          description: Invalid input.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        default:
          description: Unexpected error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - OAuth2BearerAuth: [write:properties]
  /properties/{propertyId}:
    summary: Manage a specific property.
    parameters:
      - $ref: '#/components/parameters/PropertyID'
    get:
      summary: Get a property by its ID
      tags: [Properties]
      operationId: getPropertyById
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Property not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        default:
          description: Unexpected error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - OAuth2BearerAuth: [read:properties]
    put:
      summary: Update an existing property (replaces the entire resource)
      tags: [Properties]
      operationId: updateProperty
      requestBody:
        description: Property object to update.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PropertyCreateRequest' # All fields required for replacement
      responses:
        '200':
          description: Property updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
        '400':
          description: Invalid input.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Property not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        default:
          description: Unexpected error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - OAuth2BearerAuth: [write:properties]
    patch:
      summary: Partially update an existing property
      tags: [Properties]
      operationId: patchProperty
      requestBody:
        description: Property fields to update.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PropertyUpdateRequest'
      responses:
        '200':
          description: Property partially updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
        '400':
          description: Invalid input for patch operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Property not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        default:
          description: Unexpected error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - OAuth2BearerAuth: [write:properties]
    delete:
      summary: Delete a property by its ID
      tags: [Properties]
      operationId: deleteProperty
      responses:
        '204':
          description: Property deleted successfully. No content.
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Property not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict (e.g., property has associated transactions and cannot be deleted).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        default:
          description: Unexpected error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - OAuth2BearerAuth: [write:properties]
...